# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocalMind API",
        "description": "OpenAPI documentation for the LocalMind backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Docker-served LocalMind backend"
        }
    ],
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "List the authenticated user's favorite questions",
                "operationId": "64160935c36c993b87092236b3413afa",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated favorite questions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionPaginatedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/questions/{question}/favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Toggle favorite status for a question",
                "operationId": "08d75c86cbf2b426344899b28fcc5f92",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status toggled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FavoriteToggleResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get the authenticated user's profile",
                "operationId": "66564f75ea7cb0ef41884d9c52722e34",
                "responses": {
                    "200": {
                        "description": "Authenticated profile data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile/questions": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "List questions created by the authenticated user",
                "operationId": "c96e41f5de120f85a63d73e81a594d84",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated user question list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionPaginatedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AuthorSummary": {
                "required": [
                    "id",
                    "name",
                    "initial"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lahcen"
                    },
                    "initial": {
                        "type": "string",
                        "example": "L"
                    }
                },
                "type": "object"
            },
            "QuestionMetrics": {
                "required": [
                    "responses_count",
                    "favorites_count"
                ],
                "properties": {
                    "responses_count": {
                        "type": "integer",
                        "example": 4
                    },
                    "favorites_count": {
                        "type": "integer",
                        "example": 9
                    }
                },
                "type": "object"
            },
            "ResponseResource": {
                "required": [
                    "id",
                    "content",
                    "question_id",
                    "created_at",
                    "author",
                    "is_owner"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "content": {
                        "type": "string",
                        "example": "You can find a pharmacy near the train station."
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2 minutes ago"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummary"
                    },
                    "is_owner": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "QuestionResource": {
                "required": [
                    "id",
                    "title",
                    "content",
                    "location",
                    "created_at",
                    "timestamp",
                    "author",
                    "metrics",
                    "responses"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "title": {
                        "type": "string",
                        "example": "Where can I find a late-night pharmacy?"
                    },
                    "content": {
                        "type": "string",
                        "example": "I need a pharmacy open after 10pm in Casablanca."
                    },
                    "location": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "3 hours ago"
                    },
                    "timestamp": {
                        "type": "string",
                        "example": "10 Mar 2026 21:30"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummary"
                    },
                    "metrics": {
                        "$ref": "#/components/schemas/QuestionMetrics"
                    },
                    "responses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ResponseResource"
                        }
                    }
                },
                "type": "object"
            },
            "ProfileStats": {
                "required": [
                    "questions_count",
                    "responses_count",
                    "favorites_count"
                ],
                "properties": {
                    "questions_count": {
                        "type": "integer",
                        "example": 7
                    },
                    "responses_count": {
                        "type": "integer",
                        "example": 16
                    },
                    "favorites_count": {
                        "type": "integer",
                        "example": 11
                    }
                },
                "type": "object"
            },
            "ProfileResource": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "initial",
                    "joined_at",
                    "stats",
                    "recent_questions"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lahcen"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "lahcen@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "member",
                        "enum": [
                            "admin",
                            "member"
                        ]
                    },
                    "initial": {
                        "type": "string",
                        "example": "L"
                    },
                    "joined_at": {
                        "type": "string",
                        "example": "Mar 2026"
                    },
                    "stats": {
                        "$ref": "#/components/schemas/ProfileStats"
                    },
                    "recent_questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "FavoriteToggleResponse": {
                "required": [
                    "message",
                    "is_favorited",
                    "favorites_count"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Question added to favorites !"
                    },
                    "is_favorited": {
                        "type": "boolean",
                        "example": true
                    },
                    "favorites_count": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=1",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=2",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMetaLink": {
                "required": [
                    "url",
                    "label",
                    "active"
                ],
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=1",
                        "nullable": true
                    },
                    "label": {
                        "type": "string",
                        "example": "1"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "required": [
                    "current_page",
                    "from",
                    "last_page",
                    "links",
                    "path",
                    "per_page",
                    "to",
                    "total"
                ],
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaginationMetaLink"
                        }
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 31
                    }
                },
                "type": "object"
            },
            "QuestionPaginatedResponse": {
                "required": [
                    "data",
                    "links",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the bearer token returned by the register or login endpoint.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}